code

+++ PBL3/DAL/Bill_DAL.cs
@@ -31,43 +31,25 @@ namespace PBL3.DAL
             {
                 DataGridView bill = new DataGridView();
                 bill.DataSource = db.Bills
-                    .Where(p => p.thanhtoan == true)
                     .Select(p => new {
                         p.idBill,
-                        p.Customer,
+                        p.idAccount,
                         p.paymenttime,
-                        p.TableFood.name,
-                        p.idAccount
+                        p.thanhtoan,
+                        p.TableFood.name
                     }).ToList();
                 return bill;
             }
         }
-        public List<Bill> getBills()
+        public DataGridView getDetailBill(string ids)
         {
             using (QuanLyQuanCafeEntities db = new QuanLyQuanCafeEntities())
             {
-                List<Bill> bill = new List<Bill>();
-                bill = db.Bills.Where(p => p.thanhtoan == true).ToList();
-                return bill;
-            }
-        }
-        public List<CBBItem> getOrderedDrinnkCBB(string ids)
-        {
-            using (QuanLyQuanCafeEntities db = new QuanLyQuanCafeEntities())
-            {
-                List<CBBItem> items = new List<CBBItem>();
-                foreach (ItemOrder d in db.ItemOrders)
-                {
-                    if (d.idBill.ToString() == ids)
-                    {
-
-                        items.Add(new CBBItem
-                        {
-                            Value = (int)d.idFood,
-                            Text = d.Food.NameFood + " (x" + d.billquantity + "):  " + (d.billquantity * d.Food.price) + "đ"
-                        });
-                    }
-                }
+                DataGridView items = new DataGridView();
+                items.DataSource = db.ItemOrders.
+                    Where(p => p.idBill.ToString() == ids).
+                    Select(p => new { p.idFood, p.Food.NameFood, p.billquantity, p.Food.price }).
+                    ToList();
                 return items;
             }
         }
@@ -87,22 +69,27 @@ namespace PBL3.DAL
             }
             return tprice;
         }
-        public void del(string id)
+        public void del(int id)
         {
             using (QuanLyQuanCafeEntities db = new QuanLyQuanCafeEntities())
             {
+                foreach (ItemOrder d in db.ItemOrders)
+                {
+                    if (d.idBill == id)
+                    {
+                        db.ItemOrders.Remove(d);
+                    }
+                }
                 Bill b = db.Bills.Find(id);
                 db.Bills.Remove(b);
                 db.SaveChanges();
             }
         }
-        
         public void edit(Bill newBill)
         {
             using (QuanLyQuanCafeEntities db = new QuanLyQuanCafeEntities())
             {
                 Bill bill = db.Bills.Find(newBill.idBill);
-                bill.Customer = newBill.Customer;
                 bill.paymenttime = newBill.paymenttime;
                 bill.thanhtoan = newBill.thanhtoan;
                 bill.idAccount = newBill.idAccount;
@@ -112,22 +99,6 @@ namespace PBL3.DAL
                     MessageBoxButtons.OK, MessageBoxIcon.Information);
             }
         }
-
-        //public DataGridView getBillInfoByIdBillInfo()
-        //{
-        //    DataGridView dgv = new DataGridView();
-        //    using (QuanLyQuanCafeEntities db = new QuanLyQuanCafeEntities())
-        //    {
-        //        var v = db.BillInfoes
-        //            .Where
-        //            .Select(x => new {x.idBillInfo, x.idBill, x.Food.NameFood, x.Food.price, x.count});
-        //        dgv.DataSource = v;
-        //    }
-        //    return dgv;
-        //}
-
-
-
 
 
         //tra ve Bill chua thanh toan
@@ -139,30 +110,11 @@ namespace PBL3.DAL
                 {
                     if (b.thanhtoan == false)
                         return b;
+
                 }
             }
             return null;
         }
-        //tra ve list ItemOrder cho bill chua thanh toan
-        public List<ItemOrder> GetBillInfoByIdTable(int idtable)
-        {
-            List<ItemOrder> list = new List<ItemOrder>();
-            Bill bill = getBillTableByIdTable(idtable);
-            if(bill != null)
-            {
-                using (QuanLyQuanCafeEntities db = new QuanLyQuanCafeEntities())
-                {
-                    foreach (ItemOrder bi in db.ItemOrders)
-                    {
-                        if (bi.idBill == bill.idBill)
-                            list.Add(bi);
-                    }
-                }
-
-            }
-            return list;
-        }
-
         public List<ItemOrder> GetBillInfoByIdBill(int idbill)
         {
             List<ItemOrder> list = new List<ItemOrder>();
@@ -176,17 +128,196 @@ namespace PBL3.DAL
                             list.Add(bi);
                     }
                 }
-
             }
             return list;
         }
+        public void addBill(Bill bill, List<ItemOrder> lio)
+        {
+            using (QuanLyQuanCafeEntities db = new QuanLyQuanCafeEntities())
+            {
+                db.Bills.Add(bill);
+                db.ItemOrders.AddRange(lio);
+                db.SaveChanges();
+                MessageBox.Show("Đã thêm hóa đơn thành công", "Thông báo",
+                    MessageBoxButtons.OK, MessageBoxIcon.Information);
+            }
+        }
+        public Bill getBillById(int id)
+        {
+            using (QuanLyQuanCafeEntities db = new QuanLyQuanCafeEntities())
+            {
+                foreach (Bill b in db.Bills)
+                {
+                    if (b.idBill == id)
+                    {
+                        return b;
+                    }
+                }
+            }
+            return null;
+        }
+        //public DataGridView getBill()
+        //{
+        //    using (QuanLyQuanCafeEntities db = new QuanLyQuanCafeEntities())
+        //    {
+        //        DataGridView bill = new DataGridView();
+        //        bill.DataSource = db.Bills
+        //            .Where(p => p.thanhtoan == true)
+        //            .Select(p => new {
+        //                p.idBill,
+        //                p.paymenttime,
+        //                p.TableFood.name,
+        //                p.idAccount
+        //            }).ToList();
+        //        return bill;
+        //    }
+        //}
+        //public List<Bill> getBillsPaid()
+        //{
+        //    using (QuanLyQuanCafeEntities db = new QuanLyQuanCafeEntities())
+        //    {
+        //        List<Bill> bill = new List<Bill>();
+        //        bill = db.Bills.Where(p => p.thanhtoan == true).ToList();
+        //        return bill;
+        //    }
+        //}
+        public List<Bill> getBills()
+        {
+            using (QuanLyQuanCafeEntities db = new QuanLyQuanCafeEntities())
+            {
+                List<Bill> bill = new List<Bill>();
+                bill = db.Bills.ToList();
+                return bill;
+            }
+        }
+        //public List<CBBItem> getOrderedDrinnkCBB(string ids)
+        //{
+        //    using (QuanLyQuanCafeEntities db = new QuanLyQuanCafeEntities())
+        //    {
+        //        List<CBBItem> items = new List<CBBItem>();
+        //        foreach (ItemOrder d in db.ItemOrders)
+        //        {
+        //            if (d.idBill.ToString() == ids)
+        //            {
+
+        //                items.Add(new CBBItem
+        //                {
+        //                    Value = (int)d.idFood,
+        //                    Text = d.Food.NameFood + " (x" + d.billquantity + "):  " + (d.billquantity * d.Food.price) + "đ"
+        //                });
+        //            }
+        //        }
+        //        return items;
+        //    }
+        //}
+        //public int getTotal(string s)
+        //{
+        //    int tprice = 0;
+        //    using (QuanLyQuanCafeEntities db = new QuanLyQuanCafeEntities())
+        //    {
+        //        foreach (ItemOrder d in db.ItemOrders)
+        //        {
+        //            if (d.idBill.ToString() == s)
+        //            {
+
+        //                tprice += (int)(d.billquantity * d.Food.price);
+        //            }
+        //        }
+        //    }
+        //    return tprice;
+        //}
+        //public void del(string id)
+        //{
+        //    using (QuanLyQuanCafeEntities db = new QuanLyQuanCafeEntities())
+        //    {
+        //        Bill b = db.Bills.Find(id);
+        //        db.Bills.Remove(b);
+        //        db.SaveChanges();
+        //    }
+        //}
+
+        //public void edit(Bill newBill)
+        //{
+        //    using (QuanLyQuanCafeEntities db = new QuanLyQuanCafeEntities())
+        //    {
+        //        Bill bill = db.Bills.Find(newBill.idBill);
+        //        bill.paymenttime = newBill.paymenttime;
+        //        bill.thanhtoan = newBill.thanhtoan;
+        //        bill.idAccount = newBill.idAccount;
+        //        bill.idTable = newBill.idTable;
+        //        db.SaveChanges();
+        //    }
+        //}
+        //public DataGridView getDetailBill(string ids)
+        //{
+        //    using (QuanLyQuanCafeEntities db = new QuanLyQuanCafeEntities())
+        //    {
+        //        DataGridView items = new DataGridView();
+        //        items.DataSource = db.ItemOrders.
+        //            Where(p => p.idBill.ToString() == ids).
+        //            Select(p => new { p.idFood, p.Food.NameFood, p.billquantity, p.Food.price }).
+        //            ToList();
+        //        return items;
+        //    }
+        //}
+
+
+        ////public DataGridView getBillInfoByIdBillInfo()
+        ////{
+        ////    DataGridView dgv = new DataGridView();
+        ////    using (QuanLyQuanCafeEntities db = new QuanLyQuanCafeEntities())
+        ////    {
+        ////        var v = db.BillInfoes
+        ////            .Where
+        ////            .Select(x => new {x.idBillInfo, x.idBill, x.Food.NameFood, x.Food.price, x.count});
+        ////        dgv.DataSource = v;
+        ////    }
+        ////    return dgv;
+        ////}
+
+
+
+
+
+        ////tra ve Bill chua thanh toan
+        //public Bill getBillTableByIdTable(int idtable)
+        //{
+        //    using (QuanLyQuanCafeEntities db = new QuanLyQuanCafeEntities())
+        //    {
+        //        foreach (Bill b in db.Bills.Where(p => p.idTable == idtable).ToList())
+        //        {
+        //            if (b.thanhtoan == false)
+        //                return b;
+        //        }
+        //    }
+        //    return null;
+        //}
+
+        //public List<ItemOrder> GetBillInfoByIdBill(int idbill)
+        //{
+        //    List<ItemOrder> list = new List<ItemOrder>();
+        //    if (idbill.ToString() != null)
+        //    {
+        //        using (QuanLyQuanCafeEntities db = new QuanLyQuanCafeEntities())
+        //        {
+        //            foreach (ItemOrder bi in db.ItemOrders)
+        //            {
+        //                if (bi.idBill == idbill)
+        //                    list.Add(bi);
+        //            }
+        //        }
+
+        //    }
+        //    return list;
+        //}
+
 
 
 
 
         public void delItemOrder(string idbill)
         {
-            if(!string.IsNullOrEmpty(idbill))
+            if (!string.IsNullOrEmpty(idbill))
             {
                 int id = Convert.ToInt32(idbill);
                 QuanLyQuanCafeEntities db = new QuanLyQuanCafeEntities();
@@ -224,8 +355,10 @@ namespace PBL3.DAL
                 int id = Convert.ToInt32(idbill);
                 QuanLyQuanCafeEntities db = new QuanLyQuanCafeEntities();
                 var item = db.ItemOrders.Where(p => p.idBill == id).ToList();
-                db.ItemOrders.RemoveRange(item);
-
+                if (item != null)
+                {
+                    db.ItemOrders.RemoveRange(item);
+                }
                 var bill = db.Bills.FirstOrDefault(b => b.idBill == id);
                 db.Bills.Remove(bill);
